requirements.txt
Flask~=2.2.2
requests~=2.28.1
PyYAML~=6.0
beautifulsoup4~=4.11.1
Unidecode~=1.3.6
redis~=4.3.4
setuptools~=57.0.0